Extracted default header loading to external script
