+++ lib/taint/_contentTypeDetection.js
