+++ node.gyp
@@ -29,7 +29,8 @@
       'lib/taint/taint-node.js',
       'lib/taint/logger.js',
       'lib/taint/_mongodb-core.js',
-      'lib/_contentTypeDetection.js',
+      'lib/taint/_contentTypeDetection.js',
+      'lib/taint/_securityHeaderOptions.js',
       'lib/_pathTraversalCheck.js',
       'lib/taint/_pg.js',
       'lib/taint/pg_lexer.js',
